Django Backend/requirements.txt
channels==3.0.3
djangorestframework_simplejwt==4.7.1
asgiref==3.3.4
Django==3.0.7
djangorestframework==3.12.4
